backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
blinker==1.4
certifi==2023.11.17
# chardet==5.2.0
# command-not-found==0.3
cryptography==3.4.8
cssselect2==0.7.0
# dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
# distro-info==1.1+ubuntu0.1
Django==5.0.2
django-cors-headers==4.3.1
django-filer==3.1.1
django-filter==23.5
django-pika-pubsub==0.6
django-polymorphic==3.1.0
django-reset-migrations==0.4.0
django-sendsms==0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
drfaddons==0.1.4
easy-thumbnails==2.8.5
filelock==3.13.1
gunicorn==21.2.0
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
more-itertools==8.10.0
# netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pika==1.3.2
Pillow==10.0.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg==3.1.16
# psycopg2-binary==2.9.7
# PyGObject==3.42.1
PyJWT==1.7.1
pyparsing==2.4.7
# python-apt==2.4.0+ubuntu2
pytz==2023.3.post1
# PyYAML==5.4.1
referencing==0.33.0
reportlab==4.1.0
rpds-py==0.18.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
# svglib==1.5.1
# systemd-python==234
tinycss2==1.2.1
typing_extensions==4.9.0
# tzdata==2023.4
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
uritemplate==4.1.1
virtualenv==20.25.0
virtualenv-clone==0.3.0
wadllib==1.3.6
webencodings==0.5.1
zipp==1.0.0
django-model-utils==4.5.1